added ML functionality to main also converted some data types from int to float     in: weightsManager.cpp and .h successfully trains model, but requires testing to verify that learns effectively possibility for future update:     propogate more than one of best from each gen?

+++ weightsManager.cpp
@@ -11,7 +11,7 @@ void randomWeights(float weightsLayer[9][9]) {
     }
 }
 
-void boardToInputLayer(const string matrix[9], int inputLayer[9]){
+void boardToInputLayer(const string matrix[9], float inputLayer[9]){
     for (int i = 0; i < 9; i++) {
         if (matrix[i] == "x") {
             inputLayer[i] = 1;
@@ -23,7 +23,7 @@ void boardToInputLayer(const string matrix[9], int inputLayer[9]){
     }
 }
 
-void passThroughLayer(const int inputLayer[9], const float weightsLayer[9][9], float outputLayer[9]){
+void passThroughLayer(const float inputLayer[9], const float weightsLayer[9][9], float outputLayer[9]){
     for (int i = 0; i < 9; i++) {
         outputLayer[i] = 0;
         for (int j = 0; j < 9; j++) {
